Extract a TabManager interface
